requirements.txt
pdfplumber
requests
beautifulsoup4
google-generativeai
